PORT OUT de  R,MDP,P,CDP

+++ src/main/java/com/example/finca_hexagonal/domain/ports/in/pagos/DeletePagoUseCase.java
@@ -3,7 +3,7 @@ package com.example.finca_hexagonal.domain.ports.in.pagos;
 import com.example.finca_hexagonal.domain.models.Pago;
 
 public interface DeletePagoUseCase {
-    Boolean delete (Integer id);
+    Boolean delete (Long id);
 
     Pago logicalDeletion (Pago pago);
