+++ src/main/java/com/example/finca_hexagonal/domain/ports/in/pagos/FindPagosUseCase.java
@@ -1,5 +1,6 @@
 package com.example.finca_hexagonal.domain.ports.in.pagos;
 
+import com.example.finca_hexagonal.domain.models.Pago;
 import org.springframework.data.domain.Page;
 
 import java.awt.print.Pageable;
@@ -9,7 +10,7 @@ import java.util.Optional;
 public interface FindPagosUseCase {
     Page<Pago> getAll(Pageable pageable);
 
-    Optional<Pago> getById(Integer id);
+    Optional<Pago> getById(Long id);
 
    Page<Pago> filterPagoByParams(Pageable pageable,
                                  Boolean estado_pagoBrowser,
